+++ internal/controllers/configuration/zz_generated_controllers.go
@@ -28,7 +28,6 @@ import (
 	discoveryv1 "k8s.io/api/discovery/v1"
 	netv1 "k8s.io/api/networking/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	k8stypes "k8s.io/apimachinery/pkg/types"
@@ -100,11 +99,6 @@ func (r *CoreV1ServiceReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 
 	// get the relevant object
 	obj := new(corev1.Service)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: corev1.SchemeGroupVersion.String(),
-		Kind:       "Service",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -209,11 +203,6 @@ func (r *DiscoveryV1EndpointSliceReconciler) Reconcile(ctx context.Context, req 
 
 	// get the relevant object
 	obj := new(discoveryv1.EndpointSlice)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: discoveryv1.SchemeGroupVersion.String(),
-		Kind:       "EndpointSlice",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -352,11 +341,6 @@ func (r *NetV1IngressReconciler) Reconcile(ctx context.Context, req ctrl.Request
 
 	// get the relevant object
 	obj := new(netv1.Ingress)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: netv1.SchemeGroupVersion.String(),
-		Kind:       "Ingress",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -506,11 +490,6 @@ func (r *NetV1IngressClassReconciler) Reconcile(ctx context.Context, req ctrl.Re
 
 	// get the relevant object
 	obj := new(netv1.IngressClass)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: netv1.SchemeGroupVersion.String(),
-		Kind:       "IngressClass",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -596,11 +575,6 @@ func (r *KongV1KongIngressReconciler) Reconcile(ctx context.Context, req ctrl.Re
 
 	// get the relevant object
 	obj := new(kongv1.KongIngress)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1.SchemeGroupVersion.String(),
-		Kind:       "KongIngress",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -687,11 +661,6 @@ func (r *KongV1KongPluginReconciler) Reconcile(ctx context.Context, req ctrl.Req
 
 	// get the relevant object
 	obj := new(kongv1.KongPlugin)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1.SchemeGroupVersion.String(),
-		Kind:       "KongPlugin",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -834,11 +803,6 @@ func (r *KongV1KongClusterPluginReconciler) Reconcile(ctx context.Context, req c
 
 	// get the relevant object
 	obj := new(kongv1.KongClusterPlugin)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1.SchemeGroupVersion.String(),
-		Kind:       "KongClusterPlugin",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -1015,11 +979,6 @@ func (r *KongV1KongConsumerReconciler) Reconcile(ctx context.Context, req ctrl.R
 
 	// get the relevant object
 	obj := new(kongv1.KongConsumer)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1.SchemeGroupVersion.String(),
-		Kind:       "KongConsumer",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -1207,11 +1166,6 @@ func (r *KongV1Beta1KongConsumerGroupReconciler) Reconcile(ctx context.Context, 
 
 	// get the relevant object
 	obj := new(kongv1beta1.KongConsumerGroup)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1beta1.SchemeGroupVersion.String(),
-		Kind:       "KongConsumerGroup",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -1401,11 +1355,6 @@ func (r *KongV1Beta1TCPIngressReconciler) Reconcile(ctx context.Context, req ctr
 
 	// get the relevant object
 	obj := new(kongv1beta1.TCPIngress)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1beta1.SchemeGroupVersion.String(),
-		Kind:       "TCPIngress",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -1608,11 +1557,6 @@ func (r *KongV1Beta1UDPIngressReconciler) Reconcile(ctx context.Context, req ctr
 
 	// get the relevant object
 	obj := new(kongv1beta1.UDPIngress)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1beta1.SchemeGroupVersion.String(),
-		Kind:       "UDPIngress",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {
@@ -1742,11 +1686,6 @@ func (r *KongV1Alpha1IngressClassParametersReconciler) Reconcile(ctx context.Con
 
 	// get the relevant object
 	obj := new(kongv1alpha1.IngressClassParameters)
-	// set type meta to the object
-	obj.TypeMeta = metav1.TypeMeta{
-		APIVersion: kongv1alpha1.SchemeGroupVersion.String(),
-		Kind:       "IngressClassParameters",
-	}
 
 	if err := r.Get(ctx, req.NamespacedName, obj); err != nil {
 		if apierrors.IsNotFound(err) {