+++ internal/controllers/gateway/grpcroute_controller.go
@@ -272,6 +272,7 @@ func (r *GRPCRouteReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 		// for any error other than 404, requeue
 		return ctrl.Result{}, err
 	}
+
 	debug(log, grpcroute, "processing grpcroute")
 
 	// if there's a present deletion timestamp then we need to update the proxy cache