+++ internal/controllers/gateway/httproute_controller.go
@@ -346,6 +346,7 @@ func (r *HTTPRouteReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 		// for any error other than 404, requeue
 		return ctrl.Result{}, err
 	}
+
 	debug(log, httproute, "processing httproute")
 
 	// if there's a present deletion timestamp then we need to update the proxy cache