+++ internal/controllers/gateway/tcproute_controller.go
@@ -269,6 +269,7 @@ func (r *TCPRouteReconciler) Reconcile(ctx context.Context, req ctrl.Request) (c
 		// for any error other than 404, requeue
 		return ctrl.Result{}, err
 	}
+
 	debug(log, tcproute, "processing tcproute")
 
 	// if there's a present deletion timestamp then we need to update the proxy cache